init component hierarchy and concrete class structure
