+++ component/__init__.py
@@ -1,0 +1,3 @@
+from component.action_component import ActionComponent
+from component.observer_component import ObserverComponent
+from component.component import Component
