+++ component/components/__init__.py
@@ -1,0 +1,7 @@
+from component.components.components import Component
+from component.components.camera_component import CameraComponent
+from component.components.button_component import ButtonComponent
+from component.components.nfc_component import NfcComponent
+from component.components.vibration_component import VibrationComponent
+
+__all__ =  ['CameraComponent', 'ButtonComponent', 'NfcComponent', 'VibrationComponent']
