+++ component/components/button_component.py
@@ -1,0 +1,5 @@
+from component import ObserverComponent
+
+
+class ButtonComponent(ObserverComponent):
+    pass
