+++ component/components/camera_component.py
@@ -1,0 +1,5 @@
+from component import ActionComponent
+
+
+class CameraComponent(ActionComponent):
+    pass
