+++ component/components/components.py
@@ -1,0 +1,9 @@
+from enum import Enum
+from component.components import *
+
+
+class Component(Enum):
+    Camera = CameraComponent
+    Button = ButtonComponent
+    NFC = NfcComponent
+    Vibration = VibrationComponent
