+++ component/components/nfc_component.py
@@ -1,0 +1,5 @@
+from component import ActionComponent, ObserverComponent
+
+
+class NfcComponent(ActionComponent, ObserverComponent):
+    pass
