+++ component/components/vibration_component.py
@@ -1,0 +1,5 @@
+from component import ObserverComponent
+
+
+class VibrationComponent(ObserverComponent):
+    pass
